Add feature gating
